Correcion de notebooks y agregando el PB-11 y PB-12
