--- models/baseline_dt.pkl.dvc
@@ -1,5 +1,0 @@
-outs:
-- md5: f6e1bff234ca66f3b93338930c410782
-  size: 563706
-  hash: md5
-  path: baseline_dt.pkl
