--- models/baseline_gb.pkl.dvc
@@ -1,5 +1,0 @@
-outs:
-- md5: 6f42615dd1266bc41e008ab0ecf6d368
-  size: 141737
-  hash: md5
-  path: baseline_gb.pkl
