--- models/baseline_lr.pkl.dvc
@@ -1,5 +1,0 @@
-outs:
-- md5: 7f36af729a8bd63de80684e944eedcf9
-  size: 1312
-  hash: md5
-  path: baseline_lr.pkl
