--- models/baseline_rf.pkl.dvc
@@ -1,5 +1,0 @@
-outs:
-- md5: adb5f26dec7b707649ff077535123095
-  size: 53273370
-  hash: md5
-  path: baseline_rf.pkl
